Small changes in volunteer branch and organization branch. Fix most critiques from meeting. Move widgets to their own files in components, delete unused files and widgets.
