+++ lib/components/volunteer_organization_information_items_to_deliver_extended.dart
@@ -1,0 +1,114 @@
+import 'package:flutter/material.dart';
+
+import 'package:phase1/constants.dart';
+
+class OrganizationInformation extends StatelessWidget {
+  final String orgEmail;
+
+  final String orgPhone;
+
+  final String orgAddress;
+
+  OrganizationInformation({@required this.orgEmail, @required this.orgPhone, @required this.orgAddress});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        Container(
+          decoration: BoxDecoration(color: colorScheme.background, borderRadius: BorderRadius.all(Radius.circular(20))),
+          width: MediaQuery.of(context).size.width,
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16.0),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Text(
+                  'Organization Information ',
+                  style: TextStyle(
+                    fontSize: 20,
+                  ),
+                ),
+                SizedBox(
+                  height: 10,
+                ),
+                Container(
+                  height: 5,
+                  width: 100,
+                  decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
+                ),
+                SizedBox(
+                  height: 10,
+                ),
+                Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    RichText(
+                      text: TextSpan(children: <TextSpan>[
+                        TextSpan(
+                            text: 'Email Address: ',
+                            style: TextStyle(
+                              fontSize: 18,
+                              color: colorScheme.onBackground,
+                            )),
+                        TextSpan(
+                            text: orgEmail,
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w600,
+                              color: colorScheme.onBackground,
+                            ))
+                      ]),
+                    ),
+                    SizedBox(
+                      height: 10,
+                    ),
+                    RichText(
+                      text: TextSpan(children: <TextSpan>[
+                        TextSpan(
+                            text: 'Phone Number: ',
+                            style: TextStyle(
+                              fontSize: 18,
+                              color: colorScheme.onBackground,
+                            )),
+                        TextSpan(
+                            text: orgPhone,
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w600,
+                              color: colorScheme.onBackground,
+                            ))
+                      ]),
+                    ),
+                    SizedBox(
+                      height: 10,
+                    ),
+                    RichText(
+                      text: TextSpan(children: <TextSpan>[
+                        TextSpan(
+                            text: 'Donation Location: ',
+                            style: TextStyle(
+                              fontSize: 18,
+                              color: colorScheme.onBackground,
+                            )),
+                        TextSpan(
+                            text: orgAddress,
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w600,
+                              color: colorScheme.onBackground,
+                            ))
+                      ]),
+                    ),
+                  ],
+                )
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
