+++ lib/pages/volunteer/organization_profile_page.dart
@@ -17,163 +17,153 @@ class _OrganizationProfilePageState extends State<OrganizationProfilePage> {
       title: ' ',
       helpText: 'If u don\'t know how to use this app u stupid lmao',
       body: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4),
-              child: Text(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Text(
                 '(Organization Name)',
                 style: TextStyle(fontSize: 35, fontWeight: FontWeight.w900, color: purpleAccent),
               ),
-            ),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4),
-              child: Text(
+              Text(
                 'Distance: (Distance away)',
                 style: TextStyle(fontSize: 25, fontWeight: FontWeight.w700, color: purpleAccent),
               ),
-            ),
-            SizedBox(height: 20),
-            FlatButton(
-                onPressed: () {
-//TODO
-                },
-                child: Container(
-                  width: MediaQuery.of(context).size.width,
-                  decoration: BoxDecoration(color: colorScheme.onSecondary, borderRadius: BorderRadius.all(Radius.circular(20))),
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 8.0),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: <Widget>[
-                            SizedBox(
-                              height: 10,
+              SizedBox(height: 20),
+              Container(
+                width: MediaQuery.of(context).size.width,
+                decoration: BoxDecoration(color: colorScheme.onSecondary, borderRadius: BorderRadius.all(Radius.circular(20))),
+                child: Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 8.0),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.start,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: <Widget>[
+                      Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: <Widget>[
+                          SizedBox(
+                            height: 10,
+                          ),
+                          Text(
+                            'Description',
+                            style: TextStyle(
+                              fontSize: 20,
+                              fontWeight: FontWeight.w600,
                             ),
-                            Text(
-                              'Description',
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                            SizedBox(
-                              height: 5,
-                            ),
-                            Container(
-                              height: 5,
-                              width: 100,
-                              decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
-                            ),
-                            SizedBox(
-                              height: 10,
-                            ),
-                            Text(
-                                'The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.',
-                              style: TextStyle(
-                                fontSize: 17.0,
-                                fontWeight: FontWeight.w400,
-                              ),),
-                            SizedBox(
-                              height: 10,
-                            ),
-                            RichText(
-                              text: TextSpan(children: <TextSpan>[
-                                TextSpan(
-                                    text: 'Address: ',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      color: Colors.black,
-                                    )),
-                                TextSpan(
-                                    text: 'hello',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      fontWeight: FontWeight.w600,
-                                      color: Colors.black,
-                                    ))
-                              ]),
-                            ),
-                            SizedBox(
-                              height: 10,
-                            ),
-                            RichText(
-                              text: TextSpan(children: <TextSpan>[
-                                TextSpan(
-                                    text: 'Website: ',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      color: Colors.black,
-                                    )),
-                                TextSpan(
-                                    text: 'hello',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      fontWeight: FontWeight.w600,
-                                      color: Colors.black,
-                                    ))
-                              ]),
-                            ),
-                            SizedBox(
-                              height: 10,
-                            ),
-                            RichText(
-                              text: TextSpan(children: <TextSpan>[
-                                TextSpan(
-                                    text: 'Phone Number: ',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      color: Colors.black,
-                                    )),
-                                TextSpan(
-                                    text: 'hello',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      fontWeight: FontWeight.w600,
-                                      color: Colors.black,
-                                    ))
-                              ]),
-                            ),
-                            SizedBox(
-                              height: 10,
-                            ),
-                            RichText(
-                              text: TextSpan(children: <TextSpan>[
-                                TextSpan(
-                                    text: 'Email: ',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      color: Colors.black,
-                                    )),
-                                TextSpan(
-                                    text: 'hello',
-                                    style: TextStyle(
-                                      fontSize: 17,
-                                      fontWeight: FontWeight.w600,
-                                      color: Colors.black,
-                                    ))
-                              ]),
-                            ),
-                            SizedBox(
-                              height: 10,
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
+                          ),
+                          SizedBox(
+                            height: 5,
+                          ),
+                          Container(
+                            height: 5,
+                            width: 100,
+                            decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
+                          ),
+                          SizedBox(
+                            height: 10,
+                          ),
+                          Text(
+                              'The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.',
+                            style: TextStyle(
+                              fontSize: 17.0,
+                              fontWeight: FontWeight.w400,
+                            ),),
+                          SizedBox(
+                            height: 10,
+                          ),
+                          RichText(
+                            text: TextSpan(children: <TextSpan>[
+                              TextSpan(
+                                  text: 'Address: ',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    color: Colors.black,
+                                  )),
+                              TextSpan(
+                                  text: 'hello',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    fontWeight: FontWeight.w600,
+                                    color: Colors.black,
+                                  ))
+                            ]),
+                          ),
+                          SizedBox(
+                            height: 10,
+                          ),
+                          RichText(
+                            text: TextSpan(children: <TextSpan>[
+                              TextSpan(
+                                  text: 'Website: ',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    color: Colors.black,
+                                  )),
+                              TextSpan(
+                                  text: 'hello',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    fontWeight: FontWeight.w600,
+                                    color: Colors.black,
+                                  ))
+                            ]),
+                          ),
+                          SizedBox(
+                            height: 10,
+                          ),
+                          RichText(
+                            text: TextSpan(children: <TextSpan>[
+                              TextSpan(
+                                  text: 'Phone Number: ',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    color: Colors.black,
+                                  )),
+                              TextSpan(
+                                  text: 'hello',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    fontWeight: FontWeight.w600,
+                                    color: Colors.black,
+                                  ))
+                            ]),
+                          ),
+                          SizedBox(
+                            height: 10,
+                          ),
+                          RichText(
+                            text: TextSpan(children: <TextSpan>[
+                              TextSpan(
+                                  text: 'Email: ',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    color: Colors.black,
+                                  )),
+                              TextSpan(
+                                  text: 'hello',
+                                  style: TextStyle(
+                                    fontSize: 17,
+                                    fontWeight: FontWeight.w600,
+                                    color: Colors.black,
+                                  ))
+                            ]),
+                          ),
+                          SizedBox(
+                            height: 10,
+                          ),
+                        ],
+                      ),
+                    ],
                   ),
-                )),
-            SizedBox(height: 20),
-            Column(
-              children: <Widget>[
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                  child: Container(
+                ),
+              ),
+              SizedBox(height: 20),
+              Column(
+                children: <Widget>[
+                  Container(
                     decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.all(Radius.circular(20))),
                     width: MediaQuery.of(context).size.width,
                     child: Padding(
@@ -190,7 +180,7 @@ class _OrganizationProfilePageState extends State<OrganizationProfilePage> {
                             ),
                           ),
                           SizedBox(
-                            height: 10,
+                            height: 5,
                           ),
                           Container(
                             height: 5,
@@ -198,7 +188,7 @@ class _OrganizationProfilePageState extends State<OrganizationProfilePage> {
                             decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
                           ),
                           SizedBox(
-                            height: 20,
+                            height: 10,
                           ),
                           Text('Hygiene', style: TextStyle(
                             fontSize: 17,
@@ -258,13 +248,10 @@ class _OrganizationProfilePageState extends State<OrganizationProfilePage> {
                       ),
                     ),
                   ),
-                ),
-              ],
-            ),
-            SizedBox(height: 20),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0),
-              child: Container(
+                ],
+              ),
+              SizedBox(height: 20),
+              Container(
                 width: MediaQuery.of(context).size.width,
                 child: FlatButton(
                   onPressed: () {
@@ -276,147 +263,22 @@ class _OrganizationProfilePageState extends State<OrganizationProfilePage> {
                   ),
                   color: purpleAccent,
                   child: Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 16.0, horizontal: 20),
+                    padding: const EdgeInsets.symmetric(vertical: 12.0, horizontal: 20),
                     child: Text(
                       'Continue to Requests',
                       style: TextStyle(
                           color: Colors.white,
-                          fontSize: 15
+                          fontSize: 20
                       ),
                     ),
                   ),
                 ),
               ),
-            ),
-            SizedBox(height: 20),
-          ],
+              SizedBox(height: 20),
+            ],
+          ),
         ),
       ),
-    );
-  }
-}
-
-class InfoText extends StatelessWidget {
-  final String orgEmail;
-
-  final String orgPhone;
-
-  final String orgAddress;
-
-  InfoText({this.orgEmail, this.orgPhone, this.orgAddress});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        RichText(
-          text: TextSpan(children: <TextSpan>[
-            TextSpan(
-                text: 'Email Address: ',
-                style: TextStyle(
-                  fontSize: 18,
-                  color: colorScheme.onBackground,
-                )),
-            TextSpan(
-                text: orgEmail,
-                style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.w600,
-                  color: colorScheme.onBackground,
-                ))
-          ]),
-        ),
-        FlatButton(
-            onPressed: () {
-              //TODO Redirect
-            },
-            child: Container(
-              decoration: BoxDecoration(
-                color: purpleAccent,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 6, horizontal: 20),
-                child: Text(
-                  'Open in Mail',
-                  style: TextStyle(fontSize: 15, color: colorScheme.background),
-                ),
-              ),
-            )),
-        SizedBox(
-          height: 10,
-        ),
-        RichText(
-          text: TextSpan(children: <TextSpan>[
-            TextSpan(
-                text: 'Phone Number: ',
-                style: TextStyle(
-                  fontSize: 18,
-                  color: colorScheme.onBackground,
-                )),
-            TextSpan(
-                text: orgPhone,
-                style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.w600,
-                  color: colorScheme.onBackground,
-                ))
-          ]),
-        ),
-        FlatButton(
-            onPressed: () {
-              //TODO Redirect
-            },
-            child: Container(
-              decoration: BoxDecoration(
-                color: purpleAccent,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 6, horizontal: 20),
-                child: Text(
-                  'Call',
-                  style: TextStyle(fontSize: 15, color: colorScheme.background),
-                ),
-              ),
-            )),
-        RichText(
-          text: TextSpan(children: <TextSpan>[
-            TextSpan(
-                text: 'Donation Location: ',
-                style: TextStyle(
-                  fontSize: 18,
-                  color: colorScheme.onBackground,
-                )),
-            TextSpan(
-                text: orgAddress,
-                style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.w600,
-                  color: colorScheme.onBackground,
-                ))
-          ]),
-        ),
-        FlatButton(
-            onPressed: () {
-              //TODO Redirect
-            },
-            child: Container(
-              decoration: BoxDecoration(
-                color: purpleAccent,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 6, horizontal: 20),
-                child: Text(
-                  'Open in Maps',
-                  style: TextStyle(fontSize: 15, color: colorScheme.background),
-                ),
-              ),
-            )),
-      ],
     );
   }
 }
