+++ lib/pages/volunteer/settings_page.dart
@@ -167,38 +167,37 @@ class DeleteAccount extends StatelessWidget {
       padding: EdgeInsets.symmetric(horizontal: 16.0),
       child: InkWell(
         onTap: () {
-          auth.signOut();
-//          showDialog(
-//            context: context,
-//            builder: (_) => AlertDialog(
-//              shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5.0)),
-//              title:
-//                  Text('You cannot recover your account once you delete it. All progress, past donations, and current donations will be deleted too. '
-//                      'Are you sure to complete this process?'),
-//              actions: [
-//                FlatButton(
-//                  child: Text(
-//                    'Cancel',
-//                    style: TextStyle(color: colorScheme.secondary),
-//                  ),
-//                  onPressed: () {
-//                    Navigator.pop(context);
-//                  },
-//                ),
-//                FlatButton(
-//                  child: Text(
-//                    'Delete My Account',
-//                    style: TextStyle(color: colorScheme.error),
-//                  ),
-//                  onPressed: () {
-////                    auth.currentUser().then((user) => user.delete());
-//                  auth.signOut();
-//                    //TODO Delete DB
-//                  },
-//                ),
-//              ],
-//            ),
-//          );
+          showDialog(
+            context: context,
+            builder: (_) => AlertDialog(
+              shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5.0)),
+              title:
+                  Text('You cannot recover your account once you delete it. All progress, past donations, and current donations will be deleted too. '
+                      'Are you sure to complete this process?'),
+              actions: [
+                FlatButton(
+                  child: Text(
+                    'Cancel',
+                    style: TextStyle(color: colorScheme.secondary),
+                  ),
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                ),
+                FlatButton(
+                  child: Text(
+                    'Delete My Account',
+                    style: TextStyle(color: colorScheme.error),
+                  ),
+                  onPressed: () {
+//                    auth.currentUser().then((user) => user.delete());
+                  auth.signOut();
+                    //TODO Delete DB
+                  },
+                ),
+              ],
+            ),
+          );
         },
         child: Container(
           height: 45.0,
@@ -217,107 +216,3 @@ class DeleteAccount extends StatelessWidget {
     );
   }
 }
-
-class PrivacyToggleSwitch extends StatefulWidget {
-  @override
-  _PrivacyToggleSwitchState createState() => _PrivacyToggleSwitchState();
-}
-
-class _PrivacyToggleSwitchState extends State<PrivacyToggleSwitch> {
-  bool isSwitched = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      trailing: Switch(
-        value: isSwitched,
-        onChanged: (value) {
-          setState(() {
-            isSwitched = value;
-            print(isSwitched);
-          });
-        },
-        activeTrackColor: colorScheme.onSecondary,
-        activeColor: purpleAccent,
-      ),
-      title: Text(
-        'Enable Location Tracking',
-        style: TextStyle(
-          color: colorScheme.onBackground,
-          fontSize: 18,
-        ),
-      ),
-    );
-  }
-}
-
-class NotificationToggleSwitch extends StatefulWidget {
-  @override
-  _NotificationToggleSwitchState createState() => _NotificationToggleSwitchState();
-}
-
-class _NotificationToggleSwitchState extends State<NotificationToggleSwitch> {
-  bool isSwitched = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      trailing: Switch(
-        value: isSwitched,
-        onChanged: (value) {
-          setState(() {
-            isSwitched = value;
-            print(isSwitched);
-          });
-        },
-        activeTrackColor: colorScheme.onSecondary,
-        activeColor: purpleAccent,
-      ),
-      title: Text(
-        'Enable Push Notifications',
-        style: TextStyle(
-          color: colorScheme.onBackground,
-          fontSize: 18,
-        ),
-      ),
-    );
-  }
-}
-
-//class UserPrivacy extends StatelessWidget {
-//  @override
-//  Widget build(BuildContext context) {
-//    return Padding(
-//      padding: const EdgeInsets.symmetric(horizontal: 16.0),
-//      child: Container(
-//        decoration: BoxDecoration(color: colorScheme.onSecondary, borderRadius: BorderRadius.all(Radius.circular(20))),
-//        height: 224,
-//        width: MediaQuery.of(context).size.width,
-//        child: Padding(
-//          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16.0),
-//          child: Column(
-//            mainAxisAlignment: MainAxisAlignment.start,
-//            crossAxisAlignment: CrossAxisAlignment.start,
-//            children: <Widget>[
-//              Text(
-//                'User Privacy',
-//                style: TextStyle(
-//                  fontSize: 24,
-//                ),
-//              ),
-//              SizedBox(
-//                height: 5,
-//              ),
-//              Container(
-//                height: 5,
-//                width: 100,
-//                decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
-//              ),
-//              PrivacyToggleSwitch(),
-//            ],
-//          ),
-//        ),
-//      ),
-//    );
-//  }
-//}
