+++ lib/pages/volunteer/volunteer_donate_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:phase1/components/volunteer_donate_page_item_selection.dart';
 import 'package:phase1/pages/volunteer/donation_confirmation_page.dart';
 
-import '../../components/item_increment.dart';
 import '../../components/standard_layout.dart';
 import '../../constants.dart';
 
@@ -49,12 +49,12 @@ class _VolunteerDonatePageState extends State<VolunteerDonatePage> {
                   ),
                   color: purpleAccent,
                   child: Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 16.0, horizontal: 20),
+                    padding: const EdgeInsets.symmetric(vertical: 12.0, horizontal: 20),
                     child: Text(
                       'Continue',
                       style: TextStyle(
                           color: Colors.white,
-                          fontSize: 15
+                          fontSize: 20
                       ),
                     ),
                   ),
@@ -63,58 +63,6 @@ class _VolunteerDonatePageState extends State<VolunteerDonatePage> {
               SizedBox(height: 15),
             ],
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class DonatePageItemSelection extends StatelessWidget {
-  final String boxTitle;
-
-  DonatePageItemSelection({this.boxTitle});
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(color: colorScheme.onSecondary, borderRadius: BorderRadius.all(Radius.circular(20))),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 8.0),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            SizedBox(
-              height: 10,
-            ),
-            Row(
-              children: <Widget>[
-                Text(
-                  boxTitle,
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-              ],
-            ),
-            SizedBox(
-              height: 10,
-            ),
-            Container(
-              height: 5,
-              width: 50,
-              decoration: BoxDecoration(color: purpleAccent, borderRadius: BorderRadius.circular(21)),
-            ),
-            SizedBox(
-              height: 10,
-            ),
-            ItemIncrement(itemName: 'Toilet paper'),
-            SizedBox(height: 10),
-            ItemIncrement(itemName: 'Pillow'),
-            SizedBox(height: 10),
-            ItemIncrement(itemName: 'Shoes'),
-            SizedBox(height: 10),
-          ],
         ),
       ),
     );
